added GetStocks but not implemented in main code need to add this as a filter to filter out my stocks.csv file from nifty index contituents
